requirements.txt
fastapi==0.74.1
jinja2==3.0.3
uvicorn==0.17.6
python-multipart==0.0.5
requests==2.27.1
starlette==0.17.1
requests_kerberos==0.14.0
itsdangerous==2.1.0
SQLAlchemy==1.4.31
alembic==1.7.6
gunicorn==20.1.0
asyncpg==0.25.0
marshmallow==3.13.0
marshmallow-dataclass==8.5.3
marshmallow-enum==1.5.1
pandas==1.4.2
pydantic==1.9.0
python-jose==3.3.0
backoff==1.10.0
fastapi_utils==0.2.1
thrift==0.11.0
thrift-sasl==0.4.3
JayDeBeApi==1.2.3
python-dotenv==0.21.0